// web/platform-admin-web/src/api/generated/model/announcementPageItemAudience.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 공지 대상
 */
export type AnnouncementPageItemAudience = typeof AnnouncementPageItemAudience[keyof typeof AnnouncementPageItemAudience];


export const AnnouncementPageItemAudience = {
  ALL: 'ALL',
  PLATFORM: 'PLATFORM',
  ORGANIZATION: 'ORGANIZATION',
} as const;