// web/platform-admin-web/src/api/generated/model/announcementPageItemCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 공지 분류
 */
export type AnnouncementPageItemCategory = typeof AnnouncementPageItemCategory[keyof typeof AnnouncementPageItemCategory];


export const AnnouncementPageItemCategory = {
  NOTICE: 'NOTICE',
  MAINTENANCE: 'MAINTENANCE',
  SECURITY: 'SECURITY',
  EVENT: 'EVENT',
} as const;