// web/platform-admin-web/src/api/generated/model/announcementPageItemPriority.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 공지 우선순위
 */
export type AnnouncementPageItemPriority = typeof AnnouncementPageItemPriority[keyof typeof AnnouncementPageItemPriority];


export const AnnouncementPageItemPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
} as const;