// web/platform-admin-web/src/api/generated/model/announcementPageItemStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 게시 상태
 */
export type AnnouncementPageItemStatus = typeof AnnouncementPageItemStatus[keyof typeof AnnouncementPageItemStatus];


export const AnnouncementPageItemStatus = {
  DRAFT: 'DRAFT',
  SCHEDULED: 'SCHEDULED',
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
} as const;